// ui/src/app/model/JobQueryResult.ts
/**
 * Job Monitor Service
 * Job Monitor API for interacting with asynchronous batch jobs and workflows.
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

/**
 * Result for an individual job returned by a job query
 */
export interface JobQueryResult {
    /**
     * Job ID
     */
    id: string;

    /**
     * Job name
     */
    name: string;

    /**
     * Job status
     */
    status: string;

    /**
     * Job start datetime
     */
    start: Date;

    /**
     * Job end datetime
     */
    end?: Date;

    /**
     * Custom job labels with string values
     */
    labels?: any;

}