// ui/src/app/model/JobQueryRequest.ts
/**
 * Job Monitor Service
 * Job Monitor API for interacting with asynchronous batch jobs and workflows.
 *
 * OpenAPI spec version: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

/**
 * Job query parameters
 */
export interface JobQueryRequest {
    /**
     * Returns only jobs with an equal or later start datetime.  Can be specified at most once. If both start and end date are specified, start date must be before or equal to end date. 
     */
    start?: Date;

    /**
     * Returns only jobs with an equal or earlier end datetime.  Can be specified at most once. If both start and end date are specified, start date must be before or equal to end date. 
     */
    end?: Date;

    /**
     * Returns only jobs with the specified status.  If specified multiple times, returns jobs in any of the specified statuses. 
     */
    statuses?: Array<JobQueryRequest.StatusesEnum>;

    /**
     * Returns only jobs with the specified name.  If specified multiple times, returns jobs with any of the specified names. 
     */
    name?: string;

    /**
     * Returns only jobs with the given parent ID. This parameter may be unsupported for some API implementations or required for others depending on whether there exists a logical noun above a job in the resource hierarchy (for example, a cloud project). If the presence of this parameter is incompatible with the server, it may return a 400 HTTP status. 
     */
    parentId?: string;

}
export namespace JobQueryRequest {
    export enum StatusesEnum {
        Submitted = <any> 'Submitted',
        Running = <any> 'Running',
        Aborting = <any> 'Aborting',
        Failed = <any> 'Failed',
        Succeeded = <any> 'Succeeded',
        Aborted = <any> 'Aborted'
    }
}